refactor: rename folders and delete old training scripts
