--- ViT-classification-refactored/requirements.txt
@@ -1,8 +1,0 @@
-datasets==2.14.1
-evaluate==0.4.0
-numpy==1.25.1
-picsellia==6.9.0
-pycocotools==2.0.6
-torchvision==0.15.2
-transformers==4.31.0
-scikit-learn==1.3.0
