+++ ViT-classification/requirements.txt
@@ -5,4 +5,4 @@ picsellia==6.9.0
 pycocotools==2.0.6
 torchvision==0.15.2
 transformers==4.31.0
-scikit-learn
+scikit-learn==1.3.0
