--- yolov8-classification-old/requirements.txt
@@ -1,4 +1,0 @@
-pycocotools==2.0.6
-ultralytics==8.0.67
-onnx==1.13.1
-scikit-learn==1.2.2
