+++ yolov8-detection/Dockerfile
@@ -4,13 +4,12 @@ COPY ./yolov8-detection/requirements.txt .
 
 ARG REBUILD_ALL
 RUN python3.10 -m pip install -r requirements.txt --no-cache-dir
-ARG REBUILD_PICSELLIA
-
-RUN python3.10 -m pip install picsellia --upgrade
 
 WORKDIR /experiment
 
 COPY ./yolov8-detection/experiment ./
 COPY ./evaluator ./evaluator
+COPY ./abstract_trainer ./abstract_trainer
+COPY ./core_utils ./core_utils
 
 ENTRYPOINT ["run", "main.py"]
