+++ yolov8-detection/experiment/detection_trainer.py
