+++ yolov8-detection/experiment/trainer.py
@@ -1,7 +1,7 @@
 import os
 import sys
 
-sys.path.append(os.path.join(os.getcwd(), "yolov8-detection-refactored", "experiment"))
+sys.path.append(os.path.join(os.getcwd(), "yolov8-detection", "experiment"))
 
 from picsellia.sdk.asset import MultiAsset
