+++ yolov8-detection/requirements.txt
@@ -1,3 +1,5 @@
 pycocotools==2.0.6
 ultralytics==8.0.48
 onnx==1.13.0
+coverage==7.3.0
+picsellia==6.10.3
