--- yolov8-segmentation-refactored/requirements.txt
@@ -1,5 +1,0 @@
-pycocotools==2.0.6
-ultralytics==8.0.67
-onnx==1.13.0
-coverage==7.3.0
-picsellia==6.10.3
