+++ yolov8-segmentation/requirements.txt
@@ -1,3 +1,5 @@
 pycocotools==2.0.6
 ultralytics==8.0.67
 onnx==1.13.0
+coverage==7.3.0
+picsellia==6.10.3
